backend/pyproject.toml
[project]
name = "ai-research-paper-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "requests>=2.32.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.77",
    "uvicorn>=0.37.0",
    "pydantic>=2.11.9",
    "pymongo>=4.15.2",
    "bcrypt>=5.0.0",
    "pinecone>=7.3.0",
    "langchain-google-genai>=2.1.12",
    "pip>=25.2",
    "tqdm>=4.67.1",
    "python-multipart>=0.0.20",
    "pypdf>=6.1.1",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.1.1",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.10",
    "python-jose[cryptography]>=3.5.0",
    "passlib>=1.7.4",
]
